requirements: downgrade sqlalchemy
SQLAlchemy 1.4 removed the deprecated postgres dialect name, the name postgresql must be used instead now. The dialect is the part before the :// in the URL. SQLAlchemy 1.3 and earlier showed a deprecation warning but still accepted it.
Reference: https://help.heroku.com/ZKNTJQSK/why-is-sqlalchemy-1-4-x-not-connecting-to-heroku-postgres
So downgrade to 1.3x for Heroku deploy

Change-Id: I534029e1ccc6d5e15d774f5fbe492590cecbcfbc